feat: move componentsto ui directory
