+++ src/components/molecules/LeftSidePanelContent.tsx
@@ -7,7 +7,7 @@ import { Add } from "@mui/icons-material";
 import { Socket } from "socket.io-client";
 import { AddButton } from "../ui/buttons/AddButton";
 import { AddPair } from "../../api/workflow";
-import { GenericModal } from "../atoms/GenericModal";
+import { GenericModal } from "../ui/GenericModal";
 import { PairEditForm } from "./PairEditForm";
 import { Fab, Tooltip, Typography } from "@mui/material";
